repair_agent/requirements.txt
beautifulsoup4>=4.12.2
colorama==0.4.6
distro==1.8.0
openai==0.27.8
playsound==1.2.2
python-dotenv==0.21
pyyaml==6.0
PyPDF2
python-docx
markdown
pylatexenc
readability-lxml==0.8.1
requests
tiktoken==0.3.3
gTTS==2.3.1
docker
duckduckgo-search==3.0.2
google-api-python-client #(https://developers.google.com/custom-search/v1/overview)
pinecone-client==2.2.1
redis
orjson==3.8.10
ftfy>=6.1.1
Pillow
selenium==4.1.4
webdriver-manager
jsonschema
click
charset-normalizer>=3.1.0
spacy>=3.0.0,<4.0.0
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.5.0/en_core_web_sm-3.5.0-py3-none-any.whl
prompt_toolkit>=3.0.38
pydantic
inflection
agbenchmark
agent-protocol>=0.1.1
python-dotenv
fuzzywuzzy
langchain
antlr4-python3-runtime
# web server
fastapi
uvicorn

##Dev
coverage
flake8
numpy
pre-commit
black
isort
gitpython==3.1.31
auto-gpt-plugin-template @ git+https://github.com/Significant-Gravitas/Auto-GPT-Plugin-Template@0.1.0
mkdocs
mkdocs-material
mkdocs-table-reader-plugin
pymdown-extensions
mypy
types-Markdown
types-beautifulsoup4
types-colorama
types-Pillow
# Related to defects4j experiment
whatthepatch
# OpenAI and Generic plugins import
openapi-python-client==0.13.4

# Items below this point will not be included in the Docker Image

